+++ example/lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-
 import 'package:multiutillib/multiutillib.dart';
-import 'package:multiutillib/enums/dialog_animation_type.dart';
 import 'package:connectivity_wrapper/connectivity_wrapper.dart';
 
 void main() {
@@ -244,7 +242,9 @@ class _MyHomePageState extends State<MyHomePage> with SingleTickerProviderStateM
                   margin: const EdgeInsets.only(top: 10),
                   onPressed: () async {
                     // calling show progress dialog method
-                    await showProgressDialog(context, /*progressWidget: const CircularProgressIndicator()*/);
+                    await showProgressDialog(
+                      context, /*progressWidget: const CircularProgressIndicator()*/
+                    );
 
                     await Future.delayed(const Duration(seconds: 3));