+++ lib/src/widgets/material_card.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 class MaterialCard extends StatelessWidget {
   final Color color;
   final Widget child;
-  final Function? onTap;
   final Color? shadowColor;
+  final GestureTapCallback? onTap;
   final double? elevation, borderRadius;
   final EdgeInsetsGeometry margin, padding;
   final BorderRadiusGeometry? borderRadiusGeometry;
@@ -29,7 +29,7 @@ class MaterialCard extends StatelessWidget {
     return Container(
       margin: margin,
       child: InkWell(
-        onTap: onTap as void Function()?,
+        onTap: onTap,
         child: Material(
           color: color,
           elevation: elevation!,